Updated Route class.
Added Route::get functionality which allows you to match a route with the current request uri, it also has support for reading out variables.

Also added Route::notFound method which only works correctly when placed after the Route::get checks atm. The idea is to use this method as a default/page not found check.